requirements.txt
imreg-dft==2.0.0
numpy==1.15.2
opencv-python==3.4.3.18
Pillow==5.1.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0